update with vaildation

+++ Insurance_Project/src/main/java/com/monocept/app/dto/PlanRequestDto.java
@@ -1,6 +1,8 @@
 package com.monocept.app.dto;
 
 
+import java.math.BigDecimal;
+
 import com.monocept.app.enums.PremiumType;
 
 import jakarta.validation.constraints.DecimalMin;
@@ -26,11 +28,11 @@ public class PlanRequestDto {
 
     @NotNull(message = "Coverage amount is required")
     @DecimalMin(value = "0.01", message = "Coverage amount must be greater than zero")
-    private Double coverageAmount;
+    private BigDecimal coverageAmount;
 
     @NotNull(message = "Premium amount is required")
     @DecimalMin(value = "0.01", message = "Premium amount must be greater than zero")
-    private Double premiumAmount;
+    private BigDecimal premiumAmount;
 
     @NotNull(message = "Premium type is required")
     private PremiumType premiumType;
